gateway_agent/pyproject.toml
[project]
name = "hrt-gateway"
version = "0.1.0"
description = "FastAPI gateway for HRT MVP"
requires-python = ">=3.10,<3.13"
dependencies = [
  "fastapi==0.115.0",
  "paho-mqtt==1.6.1",
  "uvicorn[standard]==0.30.6",
]

[tool.uv]
package = false
